finished splitting up KerasLayerTest
