+++ deeplearning4j-modelimport/src/test/java/org/deeplearning4j/nn/modelimport/keras/layers/pooling/KerasPooling2DTest.java
@@ -1,0 +1,84 @@
+/*-
+ *
+ *  * Copyright 2017 Skymind,Inc.
+ *  *
+ *  *    Licensed under the Apache License, Version 2.0 (the "License");
+ *  *    you may not use this file except in compliance with the License.
+ *  *    You may obtain a copy of the License at
+ *  *
+ *  *        http://www.apache.org/licenses/LICENSE-2.0
+ *  *
+ *  *    Unless required by applicable law or agreed to in writing, software
+ *  *    distributed under the License is distributed on an "AS IS" BASIS,
+ *  *    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  *    See the License for the specific language governing permissions and
+ *  *    limitations under the License.
+ *
+ */
+package org.deeplearning4j.nn.modelimport.keras.layers.pooling;
+
+import org.deeplearning4j.nn.conf.ConvolutionMode;
+import org.deeplearning4j.nn.conf.layers.PoolingType;
+import org.deeplearning4j.nn.conf.layers.SubsamplingLayer;
+import org.deeplearning4j.nn.modelimport.keras.config.Keras1LayerConfiguration;
+import org.deeplearning4j.nn.modelimport.keras.config.Keras2LayerConfiguration;
+import org.deeplearning4j.nn.modelimport.keras.config.KerasLayerConfiguration;
+import org.junit.Test;
+
+import java.util.ArrayList;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
+
+import static org.junit.Assert.assertArrayEquals;
+import static org.junit.Assert.assertEquals;
+
+/**
+ * @author Max Pumperla
+ */
+public class KerasPooling2DTest {
+
+    private final String LAYER_NAME = "test_layer";
+    private final int[] KERNEL_SIZE = new int[]{1, 2};
+    private final int[] STRIDE = new int[]{3, 4};
+    private final PoolingType POOLING_TYPE = PoolingType.MAX;
+    private final String BORDER_MODE_VALID = "valid";
+    private final int[] VALID_PADDING = new int[]{0, 0};
+
+    private Integer keras1 = 1;
+    private Integer keras2 = 2;
+    private Keras1LayerConfiguration conf1 = new Keras1LayerConfiguration();
+    private Keras2LayerConfiguration conf2 = new Keras2LayerConfiguration();
+
+    @Test
+    public void testPooling2DLayer() throws Exception {
+        buildPooling2DLayer(conf1, keras1);
+        buildPooling2DLayer(conf2, keras2);
+    }
+
+    void buildPooling2DLayer(KerasLayerConfiguration conf, Integer kerasVersion) throws Exception {
+        Map<String, Object> layerConfig = new HashMap<String, Object>();
+        layerConfig.put(conf.getLAYER_FIELD_CLASS_NAME(), conf.getLAYER_CLASS_NAME_MAX_POOLING_2D());
+        Map<String, Object> config = new HashMap<>();
+        config.put(conf.getLAYER_FIELD_NAME(), LAYER_NAME);
+        List<Integer> kernelSizeList = new ArrayList<>();
+        kernelSizeList.add(KERNEL_SIZE[0]);
+        kernelSizeList.add(KERNEL_SIZE[1]);
+        config.put(conf.getLAYER_FIELD_POOL_SIZE(), kernelSizeList);
+        List<Integer> subsampleList = new ArrayList<>();
+        subsampleList.add(STRIDE[0]);
+        subsampleList.add(STRIDE[1]);
+        config.put(conf.getLAYER_FIELD_POOL_STRIDES(), subsampleList);
+        config.put(conf.getLAYER_FIELD_BORDER_MODE(), BORDER_MODE_VALID);
+        layerConfig.put(conf.getLAYER_FIELD_CONFIG(), config);
+        layerConfig.put(conf.getLAYER_FIELD_KERAS_VERSION(), kerasVersion);
+
+        SubsamplingLayer layer = new KerasPooling2D(layerConfig).getSubsampling2DLayer();
+        assertEquals(LAYER_NAME, layer.getLayerName());
+        assertArrayEquals(KERNEL_SIZE, layer.getKernelSize());
+        assertArrayEquals(STRIDE, layer.getStride());
+        assertEquals(POOLING_TYPE, layer.getPoolingType());
+        assertEquals(ConvolutionMode.Truncate, layer.getConvolutionMode());
+        assertArrayEquals(VALID_PADDING, layer.getPadding());
+    }
+}
